+++ src/main/java/com/thomas/gestionDeStock/services/impl/CommandeClientServiceImpl.java
@@ -195,6 +195,11 @@ public class CommandeClientServiceImpl implements CommandeClientService {
             log.error("L'id de la commande client n'existe pas");
             return;
         }
+        List<LigneCommandeClient> ligneCommandeClients = ligneCommandeClientRepository.findAllByCommandeClientId(id);
+        if (ligneCommandeClients.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une commande client en cours",
+                    ErrorCodes.COMMANDE_CLIENT_ALREADY_IN_USE);
+        }
         commandeClientRepository.deleteById(id);
     }
