+++ src/main/java/com/thomas/gestionDeStock/services/impl/CommandeFournisseurServiceImpl.java
@@ -246,6 +246,11 @@ public class CommandeFournisseurServiceImpl implements CommandeFournisseurServic
             log.error("L'id de la commande fournisseur n'existe pas");
             return;
         }
+        List<LigneCommandeFournisseur> ligneCommandeFournisseurs = ligneCommandeFournisseurRepository.findAllByCommandeFournisseurId(id);
+        if (ligneCommandeFournisseurs.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une commande fournisseur en cours",
+                    ErrorCodes.COMMANDE_FOURNISSEUR_ALREADY_IN_USE);
+        }
         commandeFournisseurRepository.deleteById(id);
     }
